requirements.txt
streamlit
feedparser
newspaper3k
pandas
scikit-learn
groq